pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Network_devices_management_system"
version = "0.1.0"
description = "基于 Fastapi 构建的网络设备管理系统"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [{ name = "lijianqiao", email = "lijianqiao2906@live.com" }]
dependencies = [
    "aerich>=0.9.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=7.0.0",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "redis[hiredis]>=6.2.0",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.34.3",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py313"
exclude = [".git", "__pycache__", ".venv", ".vscode"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "W191"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.12.2",
]

[tool.aerich]
tortoise_orm = "app.db.connection.TORTOISE_ORM"
location = "./migrations"
src_folder = "." 
